CA-286130: Set Debian 9 min memory to 256 MiB
Debian 9's stated minimum RAM requirement for amd64 is 128 MiB and the
recommended is 512 MiB. This is apparently a work of fiction since the
install initrd can't even be unpacked with 128 MiB let alone starting
any processes.

Set the minimum to 256 MiB which is enough to actually complete
installation and will then mean the template default matches the 512 MiB
recommendation.