Replace a per-template user_version with a single version
The current user_version is inconsistently applied to both leaf
templates and base templates. This is confusing and makes it difficult
to apply changes that change a variety of templates. Furthermore it is
easy to apply a change but not update all user_version in all the right
places.

Keep user_version at 1 for all default templates. Remove all the
user_version parameters from the JSON templates. When updating the
templates during postinstall or firstboot, simply remove and reload all
default templates.

The only downside of this approach is that package upgrade takes a bit
longer because there are more templates to process (takes roughly 30
seconds in my tests). Still, I think this is preferable to the current
inconsistent situation.