Flip the boot order between HDD and CD
With the way that UEFI boot entries are sorted, having the CD as the
highest priority means that the installation will be started again on
the second boot. This happens only for the auto-install ISOs because
they are missing the 'press any key to start intall' prompt.
By flipping the order, the installation will still be started for
the first boot (because there is nothing on the HDD) but for the second
boot, it will boot from the HDD as desired. This change should not have
any negative affect on BIOS boot and is also how the HVM Linux templates
work.