Revert changes to boot order
This was useful for UEFI, but UEFI is not the default anymore,
so lets not change things for BIOS.
For UEFI this is useful because with the XenRT W*-x64 UEFI ISOs you
end up with an infinite install cycle otherwise as it keeps booting from
CD after a successful install and performing yet another install.
For BIOS this didn't happen (the ISO got built differently?)

We should instead release note that for experimental UEFI support
you have to manually change the boot order,
and remember to call this out in any MRGs, as it was a very frequent
mistake during early UEFI development when using XenCenter for testing.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>