CA-336730 create-guest-templates only after network-init
create-guest-template.service sometimes fails because the management
network is reconfigured whilst it is running. This appears to be because
of network-init.service. If we run create-guest-template strictly after
network-init, then hopefully this issue will go away.

It should be noted that we can get away with using systemd to determine
whether network-init has finished (rather than using
/var/lib/misc/ran-network-init) because network-init.service has type
oneshot, and will mark itself complete when the process exits.

Signed-off-by: lippirk <ben.anson@citrix.com>